back/requirements.txt
Django>=5.0,<6.0
djangorestframework
drf-spectacular
djangorestframework-simplejwt[crypto]
django-cors-headers
psycopg2-binary
celery
redis
django-filter
